MS-19: styles refactor, move formatDate() to helpers
