+++ src/Components/Interviews/Form/form.module.css
@@ -1,34 +1,32 @@
+.container {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  max-width: 1200px;
+  margin: auto;
+}
+
+.container h1 {
+  width: 100%;
+  margin-bottom: 20px;
+  padding-bottom: 15px;
+  border-bottom: 2px rgb(191, 191, 191) solid;
+  font-size: 30px;
+}
+
 .form {
   display: flex;
+  flex-direction: row;
   flex-wrap: wrap;
+  justify-content: center;
   align-items: center;
   padding: 0px 40px 40px 40px;
   margin: auto;
 }
 
-.buttonAdd {
-  display: inline-flex;
-  width: fit-content;
-  text-decoration: none;
-}
-
-.buttonGreen {
-  padding: 16px 40px;
-  margin: 15px 2px;
-  border: rgb(21, 139, 10);
-  border-radius: 30px;
-  background-color: rgb(21, 139, 10);
-  color: rgb(255, 255, 255);
-  font-weight: bold;
-  font-size: 15px;
-  cursor: pointer;
-}
-
-.buttonGreen:hover {
-  background-color: rgb(88, 88, 88);
-  color: #fff;
-  font-style: normal;
-}
 .form fieldset {
   width: 50%;
   padding: 10px 40px;
@@ -40,6 +38,7 @@
   display: block;
   padding: 10px 5px;
 }
+
 .form input,
 .form select {
   width: 100%;
@@ -47,4 +46,31 @@
   border: 1px solid rgb(167, 167, 167);
   border-radius: 30px;
   box-sizing: border-box;
+}
+
+.form input[type="checkbox"],
+.form input[type="datetime-local"]{
+  margin:0;
+}
+
+.containerButton{
+  width: 100%;
+  text-align: center;
+}
+
+.buttonGreen {
+  padding: 16px 40px;
+  margin: 15px 2px;
+  border: rgb(21, 139, 10);
+  border-radius: 30px;
+  background-color: rgb(21, 139, 10);
+  color: rgb(255, 255, 255);
+  font-weight: bold;
+  font-size: 15px;
+}
+
+.buttonGreen:hover {
+  background-color: rgb(88, 88, 88);
+  color: #fff;
+  font-style: normal;
 }
