+++ src/Components/shared/ListItem/listItem.module.css
@@ -1,4 +1,4 @@
 .actionBtn {
   display: flex;
-  justify-content: space-evenly;
+  justify-content: space-between;
 }
